PM-6441 Reorganized UserVerificationMediatorService so everything is not in the same file
