+++ src/Core/Abstractions/IUserVerificationMediatorService.cs
@@ -1,10 +1,10 @@
-﻿using Bit.Core.Services;
+﻿using Bit.Core.Utilities.Fido2;
 
 namespace Bit.Core.Abstractions
 {
     public interface IUserVerificationMediatorService
     {
-        Task<bool> VerifyUserForFido2Async(Fido2VerificationOptions options);
+        Task<bool> VerifyUserForFido2Async(Fido2UserVerificationOptions options);
         Task<(bool CanPerfom, bool IsUnlocked)> PerformOSUnlockAsync();
         Task<(bool canPerformUnlockWithPin, bool pinVerified)> VerifyPinCodeAsync();
         Task<(bool canPerformUnlockWithMasterPassword, bool mpVerified)> VerifyMasterPasswordAsync();
