+++ src/Core/Services/UserVerification/UserVerificationMediatorService.cs
@@ -2,10 +2,11 @@
 using Bit.Core.Abstractions;
 using Bit.Core.Models.Domain;
 using Bit.Core.Resources.Localization;
+using Bit.Core.Utilities.Fido2;
 using Plugin.Fingerprint;
 using FingerprintAvailability = Plugin.Fingerprint.Abstractions.FingerprintAvailability;
 
-namespace Bit.Core.Services
+namespace Bit.Core.Services.UserVerification
 {
     public class UserVerificationMediatorService : IUserVerificationMediatorService
     {
@@ -29,7 +30,7 @@ namespace Bit.Core.Services
             _userVerificationService = userVerificationService;
         }
 
-        public async Task<bool> VerifyUserForFido2Async(Fido2VerificationOptions options)
+        public async Task<bool> VerifyUserForFido2Async(Fido2UserVerificationOptions options)
         {
             // Master Password Reprompt if enabled and shouldn't be bypassed (because of TDE)
             if (options.ShouldCheckMasterPasswordReprompt && !await _passwordRepromptService.ShouldByPassMasterPasswordRepromptAsync())
@@ -104,110 +105,6 @@ namespace Bit.Core.Services
 
             var (_, isValid) = await _platformUtilsService.ShowPasswordDialogAndGetItAsync(AppResources.MasterPassword, string.Empty, _userVerificationService.VerifyMasterPasswordAsync);
             return (true, isValid);
-        }
-    }
-
-    public readonly struct Fido2VerificationOptions
-    {
-        public Fido2VerificationOptions(bool shouldCheckMasterPasswordReprompt,
-            bool isUserVerificationRequired,
-            bool hasVaultBeenUnlockedInTransaction,
-            string rpId,
-            Action onNeedUI = null)
-        {
-            ShouldCheckMasterPasswordReprompt = shouldCheckMasterPasswordReprompt;
-            IsUserVerificationRequired = isUserVerificationRequired;
-            HasVaultBeenUnlockedInTransaction = hasVaultBeenUnlockedInTransaction;
-            RpId = rpId;
-            OnNeedUI = onNeedUI;
-        }
-
-        public bool ShouldCheckMasterPasswordReprompt { get; }
-        public bool IsUserVerificationRequired { get; }
-        public bool HasVaultBeenUnlockedInTransaction { get; }
-        public string RpId { get; }
-        public Action OnNeedUI { get; }
-    }
-
-    public interface IUserVerificationServiceStrategy
-    {
-        Task<bool> VerifyUserForFido2Async(Fido2VerificationOptions options);
-    }
-
-    public class Fido2UserVerificationRequiredServiceStrategy : IUserVerificationServiceStrategy
-    {
-        private readonly IUserVerificationMediatorService _userVerificationMediatorService;
-        private readonly IPlatformUtilsService _platformUtilsService;
-
-        public Fido2UserVerificationRequiredServiceStrategy(IUserVerificationMediatorService userVerificationMediatorService,
-            IPlatformUtilsService platformUtilsService)
-        {
-            _userVerificationMediatorService = userVerificationMediatorService;
-            _platformUtilsService = platformUtilsService;
-        }
-
-        public async Task<bool> VerifyUserForFido2Async(Fido2VerificationOptions options)
-        {
-            if (options.HasVaultBeenUnlockedInTransaction)
-            {
-                return true;
-            }
-
-            options.OnNeedUI?.Invoke();
-
-            var (canPerformOSUnlock, isOSUnlocked) = await _userVerificationMediatorService.PerformOSUnlockAsync();
-            if (canPerformOSUnlock)
-            {
-                return isOSUnlocked;
-            }
-
-            var (canPerformUnlockWithPin, pinVerified) = await _userVerificationMediatorService.VerifyPinCodeAsync();
-            if (canPerformUnlockWithPin)
-            {
-                return pinVerified;
-            }
-
-            var (canPerformUnlockWithMasterPassword, mpVerified) = await _userVerificationMediatorService.VerifyMasterPasswordAsync();
-            if (canPerformUnlockWithMasterPassword)
-            {
-                return mpVerified;
-            }
-
-            // TODO: Setup PIN code. For the sake of simplicity, we're not implementing this step now and just telling the user to do it in the main app.
-
-            await _platformUtilsService.ShowDialogAsync(AppResources.VerificationRequiredForThisActionSetUpAnUnlockMethodInBitwardenToContinue,
-                string.Format(AppResources.VerificationRequiredByX, options.RpId),
-                AppResources.Ok);
-
-            return false;
-        }
-    }
-
-    public class Fido2UserVerificationPreferredServiceStrategy : IUserVerificationServiceStrategy
-    {
-        private readonly IUserVerificationMediatorService _userVerificationMediatorService;
-
-        public Fido2UserVerificationPreferredServiceStrategy(IUserVerificationMediatorService userVerificationMediatorService)
-        {
-            _userVerificationMediatorService = userVerificationMediatorService;
-        }
-
-        public async Task<bool> VerifyUserForFido2Async(Fido2VerificationOptions options)
-        {
-            if (options.HasVaultBeenUnlockedInTransaction)
-            {
-                return true;
-            }
-
-            options.OnNeedUI?.Invoke();
-
-            var (canPerformOSUnlock, isOSUnlocked) = await _userVerificationMediatorService.PerformOSUnlockAsync();
-            if (canPerformOSUnlock)
-            {
-                return isOSUnlocked;
-            }
-
-            return false;
         }
     }
 }
