+++ src/iOS.Autofill/LoginListViewController.cs
@@ -7,6 +7,7 @@ using Bit.Core.Exceptions;
 using Bit.Core.Resources.Localization;
 using Bit.Core.Services;
 using Bit.Core.Utilities;
+using Bit.Core.Utilities.Fido2;
 using Bit.iOS.Autofill.ListItems;
 using Bit.iOS.Autofill.Models;
 using Bit.iOS.Autofill.Utilities;
@@ -204,7 +205,7 @@ namespace Bit.iOS.Autofill
                 }
 
                 return await _userVerificationMediatorService.Value.VerifyUserForFido2Async(
-                    new Fido2VerificationOptions(
+                    new Fido2UserVerificationOptions(
                         false,
                         true,
                         Context.VaultUnlockedDuringThisSession,
