+++ src/iOS.Core/Utilities/iOSCoreHelpers.cs
@@ -8,6 +8,7 @@ using Bit.App.Utilities.AccountManagement;
 using Bit.Core.Abstractions;
 using Bit.Core.Resources.Localization;
 using Bit.Core.Services;
+using Bit.Core.Services.UserVerification;
 using Bit.Core.Utilities;
 using Bit.iOS.Core.Services;
 using CoreNFC;
